requirements.txt
requests
pandas
deltalake
python-dotenv
pyarrow
matplotlib